refactor(credentials): move token adapters under credential ownership
